+++ examples/dynamic-resources/basic-with-preset-id/main.tf
@@ -1,0 +1,4 @@
+resource "mock_dynamic_resource" "example" {
+  id       = "example"
+  my_value = 0
+}