+++ examples/dynamic-resources/basic/main.tf
@@ -1,0 +1,3 @@
+resource "mock_dynamic_resource" "example" {
+  my_value = 0
+}