+++ examples/dynamic-resources/computed/main.tf
@@ -1,0 +1,1 @@
+resource "mock_dynamic_resource" "example" {}