+++ examples/dynamic-resources/multiple-resources/main.tf
@@ -1,0 +1,7 @@
+resource "mock_string_resource" "example" {
+  my_value = "Hello, world!"
+}
+
+resource "mock_integer_resource" "example" {
+  my_value = 0
+}