+++ examples/dynamic-resources/with-blocks/main.tf
@@ -1,0 +1,9 @@
+resource "mock_dynamic_resource" "example" {
+  my_values {
+    my_value = "Hello, "
+  }
+
+  my_values {
+    my_value = "world!"
+  }
+}