+++ examples/dynamic-resources/with-complex-attributes/main.tf
@@ -1,0 +1,6 @@
+resource "mock_dynamic_resource" "example" {
+  my_values = [
+    "Hello, ",
+    "World!"
+  ]
+}