+++ examples/provider/provider.tf
@@ -1,3 +1,5 @@
-provider "scaffolding" {
-  # example configuration here
+provider "mock" {
+  resource_directory = "terraform.resource"
+  data_directory     = "terraform.data"
+  use_only_state     = false
 }