+++ internal/provider/testdata/dynamic_datasource/delete/main.tf
@@ -1,0 +1,5 @@
+provider "mock" {}
+
+data "mock_simple_resource" "simple_resource" {
+  id = "simple_resource"
+}