+++ internal/provider/testdata/simple_datasource/get/main.tf
@@ -1,5 +1,9 @@
 provider mock {}
 
-data "mock_simple_resource" "test" {
+data "mock_simple_resource" "data" {
   id = "simple_resource"
+}
+
+resource "mock_simple_resource" "test" {
+  integer = data.mock_simple_resource.data.integer
 }